JetStore: Add typeahead component for file mapping, fix table selection issue #1077
