+++ jetsclient/lib/modules/form_config_impl.dart
@@ -9,6 +9,7 @@ import 'package:jetsclient/modules/user_flows/client_registry/form_config.dart';
 import 'package:jetsclient/modules/user_flows/configure_files/form_config.dart';
 import 'package:jetsclient/modules/user_flows/file_mapping/form_config.dart';
 import 'package:jetsclient/modules/user_flows/load_files/form_config.dart';
+import 'package:jetsclient/modules/user_flows/pipeline_config/form_action_delegates.dart';
 import 'package:jetsclient/modules/user_flows/pipeline_config/form_config.dart';
 import 'package:jetsclient/modules/user_flows/start_pipeline/form_config.dart';
 import 'package:jetsclient/modules/user_flows/workspace_pull/form_config.dart';
@@ -640,74 +641,6 @@ final Map<String, FormConfig> _formConfigurations = {
     formActionsDelegate: loadAllFilesActions,
   ),
 
-  // loadRawRows - Dialog to load / replace process mapping
-  FormKeys.loadRawRows: FormConfig(
-    key: FormKeys.loadRawRows,
-    title: "File Mapping Intake",
-    useListView: true,
-    actions: [
-      FormActionConfig(
-          key: ActionKeys.loadRawRowsOk,
-          capability: "client_config",
-          label: "Save",
-          buttonStyle: ActionStyle.primary,
-          leftMargin: defaultPadding,
-          rightMargin: betweenTheButtonsPadding),
-      FormActionConfig(
-          key: ActionKeys.dialogCancel,
-          label: "Cancel",
-          buttonStyle: ActionStyle.secondary,
-          leftMargin: betweenTheButtonsPadding,
-          rightMargin: defaultPadding),
-    ],
-    inputFields: [
-      [
-        // Instruction
-        TextFieldConfig(
-            label: "Enter the File Mapping Definition as csv/tsv-encoded text.",
-            maxLines: 3,
-            topMargin: defaultPadding,
-            bottomMargin: defaultPadding)
-      ],
-      [
-        FormInputFieldConfig(
-            key: FSK.rawRows,
-            label: "File Mapping (csv/tsv)",
-            hint: "Paste from spreadsheet using JetStore template",
-            flex: 1,
-            autofocus: false,
-            obscureText: false,
-            textRestriction: TextRestriction.none,
-            maxLines: 20,
-            maxLength: 51200),
-      ],
-    ],
-    formValidatorDelegate: processInputFormValidator,
-    formActionsDelegate: processInputFormActions,
-  ),
-  // Process Input Form (table as actionless form)
-  // Define ProcessInput and mapping definition
-  FormKeys.inputSourceMapping: FormConfig(
-    key: FormKeys.inputSourceMapping,
-    actions: [
-      // Action-less form
-    ],
-    inputFields: [
-      [
-        FormDataTableFieldConfig(
-            key: DTKeys.inputSourceMapping,
-            dataTableConfig: DTKeys.inputSourceMapping)
-      ],
-      [
-        FormDataTableFieldConfig(
-            key: DTKeys.processMappingTable,
-            dataTableConfig: DTKeys.processMappingTable,
-            tableHeight: double.infinity)
-      ],
-    ],
-    formValidatorDelegate: processInputFormValidator,
-    formActionsDelegate: processInputFormActions,
-  ),
   // Process Input Form (table as actionless form)
   // Define ProcessInput Configuration
   FormKeys.processInput: FormConfig(
@@ -723,7 +656,7 @@ final Map<String, FormConfig> _formConfigurations = {
             tableHeight: 800)
       ],
     ],
-    formValidatorDelegate: processInputFormValidator,
+    formValidatorDelegate: alwaysValidForm,
     formActionsDelegate: processInputFormActions,
   ),
   // addProcessInput - Dialog to add process input
@@ -827,176 +760,7 @@ final Map<String, FormConfig> _formConfigurations = {
       ],
     ],
     useListView: true,
-    formValidatorDelegate: processInputFormValidator,
-    formActionsDelegate: processInputFormActions,
-  ),
-  // processMapping - Dialog to mapping intake file structure to canonical model
-  FormKeys.processMapping: FormConfig(
-    key: FormKeys.processMapping,
-    title: "File Mapping Worksheet",
-    actions: [
-      FormActionConfig(
-          key: ActionKeys.mapperOk,
-          capability: "client_config",
-          label: "Save",
-          enableOnlyWhenFormValid: true,
-          buttonStyle: ActionStyle.primary,
-          leftMargin: defaultPadding,
-          rightMargin: betweenTheButtonsPadding,
-          bottomMargin: defaultPadding),
-      FormActionConfig(
-          key: ActionKeys.mapperDraft,
-          capability: "client_config",
-          label: "Save as Draft",
-          enableOnlyWhenFormNotValid: true,
-          buttonStyle: ActionStyle.primary,
-          leftMargin: defaultPadding,
-          rightMargin: betweenTheButtonsPadding,
-          bottomMargin: defaultPadding),
-      FormActionConfig(
-          key: ActionKeys.dialogCancel,
-          label: "Cancel",
-          buttonStyle: ActionStyle.secondary,
-          leftMargin: betweenTheButtonsPadding,
-          rightMargin: defaultPadding,
-          bottomMargin: defaultPadding),
-    ],
-    queries: {
-      "inputFieldsQuery":
-          "SELECT md.data_property, md.is_required, pm.input_column, pm.function_name, pm.argument, pm.default_value, pm.error_message FROM jetsapi.object_type_mapping_details md LEFT JOIN (SELECT * FROM jetsapi.process_mapping WHERE table_name = '{table_name}') pm ON md.data_property = pm.data_property WHERE md.object_type = '{object_type}' ORDER BY md.data_property ASC LIMIT 1000",
-      "inputColumnsQuery":
-          "SELECT column_name FROM information_schema.columns WHERE table_schema = 'public' AND table_name = '{table_name}' AND column_name NOT IN ('file_key','last_update','session_id','shard_id') ORDER BY column_name",
-      "mappingFunctionsQuery":
-          "SELECT function_name, is_argument_required FROM jetsapi.mapping_function_registry ORDER BY function_name ASC LIMIT 500",
-    },
-    inputFieldsQuery: "inputFieldsQuery",
-    savedStateQuery: "inputFieldsQuery",
-    dropdownItemsQueries: {
-      FSK.mappingFunctionsDropdownItemsCache: "mappingFunctionsQuery",
-    },
-    typeaheadItemsQueries: {
-      FSK.inputColumnsDropdownItemsCache: "inputColumnsQuery",
-    },
-    metadataQueries: {
-      FSK.mappingFunctionDetailsCache: "mappingFunctionsQuery",
-      FSK.inputColumnsCache: "inputColumnsQuery",
-    },
-    stateKeyPredicates: [FSK.objectType, FSK.tableName],
-    inputFieldRowBuilder: (index, inputFieldRow, formState) {
-      assert(inputFieldRow != null, 'error inputFieldRow should not be null');
-      if (inputFieldRow == null) {
-        return [];
-      }
-      // savedState is List<String?>? with values as per savedStateQuery
-      final savedState = formState.getCacheValue(FSK.savedStateCache) as List?;
-      final isRequired = inputFieldRow[1]! == '1';
-      final isRequiredIndicator = isRequired ? '*' : '';
-      final savedInputColumn = savedState?[index][2];
-      final inputColumnList =
-          formState.getCacheValue(FSK.inputColumnsCache) as List;
-      final inputColumnDefault =
-          inputColumnList.contains(inputFieldRow[0]) ? inputFieldRow[0] : null;
-      if (isRequired) formState.setValue(index, FSK.isRequiredFlag, "1");
-      // set the default values to the formState
-      formState.setValue(index, FSK.dataProperty, inputFieldRow[0]);
-      formState.setValue(
-          index, FSK.inputColumn, savedInputColumn ?? inputColumnDefault);
-      formState.setValue(index, FSK.functionName, savedState?[index][3]);
-      formState.setValue(index, FSK.functionArgument, savedState?[index][4]);
-      formState.setValue(index, FSK.mappingDefaultValue, savedState?[index][5]);
-      formState.setValue(index, FSK.mappingErrorMessage, savedState?[index][6]);
-      // print("Form BUILDER savedState row ${savedState![index]}");
-      return [
-        [
-          // data_property
-          TextFieldConfig(
-              label: "$index: ${inputFieldRow[0]}$isRequiredIndicator",
-              group: index,
-              flex: 1,
-              topMargin: 20.0)
-        ],
-        [
-          // input_column
-          FormTypeaheadFieldConfig(
-            key: FSK.inputColumn,
-            group: index,
-            flex: 2,
-            autovalidateMode: AutovalidateMode.always,
-            typeaheadMenuItemCacheKey: FSK.inputColumnsDropdownItemsCache,
-            defaultItem: savedInputColumn ?? inputColumnDefault,
-            inputFieldConfig: FormInputFieldConfig(
-              key: FSK.inputColumn,
-              group: index,
-              label: 'Input Column',
-              hint: 'Input File Column Name',
-              autofocus: true,
-              autovalidateMode: AutovalidateMode.always,
-              textRestriction: TextRestriction.none,
-              defaultValue: savedInputColumn ?? inputColumnDefault,
-              maxLength: 120,
-            )
-          ),
-          // FormDropdownWithSharedItemsFieldConfig(
-          //   key: FSK.inputColumn,
-          //   group: index,
-          //   flex: 2,
-          //   autovalidateMode: AutovalidateMode.always,
-          //   dropdownMenuItemCacheKey: FSK.inputColumnsDropdownItemsCache,
-          //   defaultItem: savedInputColumn ?? inputColumnDefault,
-          // ),
-          // function_name
-          FormDropdownWithSharedItemsFieldConfig(
-            key: FSK.functionName,
-            group: index,
-            flex: 1,
-            dropdownMenuItemCacheKey: FSK.mappingFunctionsDropdownItemsCache,
-            defaultItem: savedState?[index][3],
-          ),
-          // argument
-          FormInputFieldConfig(
-            key: FSK.functionArgument,
-            label: "Function Argument",
-            hint:
-                "Cleansing function argument, it is either required or ignored",
-            group: index,
-            flex: 1,
-            autovalidateMode: AutovalidateMode.always,
-            autofocus: false,
-            obscureText: false,
-            textRestriction: TextRestriction.none,
-            maxLength: 512,
-          ),
-          // default_value
-          FormInputFieldConfig(
-            key: FSK.mappingDefaultValue,
-            label: "Default Value",
-            hint:
-                "Default value to use if input value is not provided or cleansing function returns null",
-            group: index,
-            flex: 1,
-            autovalidateMode: AutovalidateMode.always,
-            autofocus: false,
-            obscureText: false,
-            textRestriction: TextRestriction.none,
-            maxLength: 512,
-          ),
-          // error_message
-          FormInputFieldConfig(
-            key: FSK.mappingErrorMessage,
-            label: "Error Message",
-            hint:
-                "Error message to raise if input value is not provided or cleansing function returns null and there is no default value",
-            group: index,
-            flex: 1,
-            autofocus: false,
-            obscureText: false,
-            textRestriction: TextRestriction.none,
-            maxLength: 125,
-          ),
-        ],
-      ];
-    },
-    formValidatorDelegate: processInputFormValidator,
+    formValidatorDelegate: pipelineConfigFormValidatorUF,
     formActionsDelegate: processInputFormActions,
   ),
