+++ jetsclient/lib/modules/user_flows/client_registry/data_table_config.dart
@@ -43,6 +43,7 @@ final Map<String, TableConfig> _tableConfigurations = {
       sortColumnName: 'option_order',
       sortAscending: true,
       noFooter: true,
+      noCopy2Clipboard: true,
       rowsPerPage: 1000000),
 
   // Client Table for Selecting Client
