+++ jetsclient/lib/modules/user_flows/client_registry/form_config.dart
@@ -137,7 +137,7 @@ final Map<String, FormConfig> _formConfigurations = {
           leftMargin: defaultPadding,
           rightMargin: betweenTheButtonsPadding),
       FormActionConfig(
-          key: ActionKeys.ufContinueLater,
+          key: ActionKeys.ufCancel,
           label: "Cancel",
           buttonStyle: ActionStyle.ufPrimary,
           leftMargin: betweenTheButtonsPadding,
