+++ jetsclient/lib/modules/user_flows/configure_files/form_action_delegates.dart
@@ -15,7 +15,7 @@ String? configureFilesFormValidator(
   switch (key) {
     case FSK.scAddOrEditSourceConfigOption:
     case FSK.scCsvOrFixedOption:
-      if (v != null) {
+      if (unpack(v) != null) {
         return null;
       }
       return "An option must be selected.";
