+++ jetsclient/lib/modules/user_flows/pipeline_config/data_table_config.dart
@@ -45,6 +45,7 @@ final Map<String, TableConfig> _tableConfigurations = {
       sortColumnName: 'option_order',
       sortAscending: true,
       noFooter: true,
+      noCopy2Clipboard: true,
       rowsPerPage: 1000000),
 
   // Pipeline Config Data Table for Pipeline Config Forms
