+++ jetsclient/lib/modules/user_flows/pipeline_config/form_action_delegates.dart
@@ -18,20 +18,22 @@ String? pipelineConfigFormValidatorUF(
     case FSK.pcAddOrEditPipelineConfigOption:
     case DTKeys.pcPipelineConfigTable:
     case FSK.mainProcessInputKey:
-    case DTKeys.pcMergedProcessInputKeys:
-    case DTKeys.pcInjectedProcessInputKeys:
     case FSK.client:
     case FSK.processName:
     case FSK.sourcePeriodType:
     case FSK.automated:
+    case DTKeys.pcProcessInputRegistry:
     case DTKeys.pcMainProcessInputKey:
     case DTKeys.pcProcessInputRegistry4MI:
-      if (v != null) {
+    case DTKeys.pcMergedProcessInputKeys:
+    case DTKeys.pcInjectedProcessInputKeys:
+      if (unpack(v) != null) {
         return null;
       }
       return "Please select an option.";
 
     case FSK.ruleConfigJson:
+    case FSK.lookbackPeriods:
       if (v != null) {
         return null;
       }
@@ -235,8 +237,10 @@ Future<String?> pipelineConfigFormActionsUF(
     // Prepare for the summary page
     case ActionKeys.pcPrepareSummaryUF:
       final main = unpackToList(state[FSK.mainProcessInputKey]);
-      final merged = unpackToList(state[FSK.mergedProcessInputKeys])??const[];
-      final injected = unpackToList(state[FSK.injectedProcessInputKeys])??const[];
+      final merged =
+          unpackToList(state[FSK.mergedProcessInputKeys]) ?? const [];
+      final injected =
+          unpackToList(state[FSK.injectedProcessInputKeys]) ?? const [];
       if (main == null) {
         print("### ufAllProcessInputKeys: $main + $merged + $injected");
         print("Error got a null main list!");
@@ -295,10 +299,10 @@ Future<String?> pipelineConfigFormActionsUF(
         final statusCode = await postSimpleAction(
             context, formState, ServerEPs.dataTableEP, encodedJsonBody);
         if (statusCode == 200) return null;
-        if(statusCode == 409 && context.mounted) {
+        if (statusCode == 409 && context.mounted) {
           showAlertDialog(context, "Record already exist, please verify.");
         }
-        if(context.mounted) {
+        if (context.mounted) {
           showAlertDialog(context, "Server error, please try again.");
         }
         return "Error while saving pipeline config";
