+++ jetsclient/lib/modules/user_flows/workspace_pull/data_table_config.dart
@@ -52,6 +52,7 @@ final Map<String, TableConfig> _tableConfigurations = {
       sortColumnName: 'option_order',
       sortAscending: true,
       noFooter: true,
+      noCopy2Clipboard: true,
       rowsPerPage: 1000000),
   // Confirm choice table
   DTKeys.wpPullWorkspaceConfirmOptions: TableConfig(
@@ -99,6 +100,7 @@ final Map<String, TableConfig> _tableConfigurations = {
       sortColumnName: 'option_order',
       sortAscending: true,
       noFooter: true,
+      noCopy2Clipboard: true,
       rowsPerPage: 1000000),
 };
